Move read_clinical_file function to utils.py
